requirements.txt
python-dotenv>=1.0.0
streamlit>=1.28.0
geopandas>=0.14.0
pandas>=2.0.0
numpy>=1.24.0
folium>=0.14.0
shapely>=2.0.0
geopy>=2.4.0
scikit-learn>=1.3.0
boto3>=1.28.0
